Fix int vs int32 type mismatch in K8s reporter tests
The VirtualMCPServerStatus.BackendCount CRD field is int32,
but test assertions compared it against int values, causing
testify's assert.Equal to fail on type mismatch.

Update the test struct's backendCount field to int32 and use
int32 literals in direct assertions to match the CRD type.

Fixes #4544

+++ cmd/thv-operator/pkg/registryapi/pgpass_test.go
@@ -277,7 +277,7 @@ func baseMCPRegistry(t *testing.T, opts ...func(*mcpv1alpha1.MCPRegistry)) *mcpv
 }
 
 func withPort(port int) func(*mcpv1alpha1.MCPRegistry) {
-	return func(r *mcpv1alpha1.MCPRegistry) { r.Spec.DatabaseConfig.Port = port }
+	return func(r *mcpv1alpha1.MCPRegistry) { r.Spec.DatabaseConfig.Port = int32(port) }
 }
 
 func withHost(host string) func(*mcpv1alpha1.MCPRegistry) {
